Firmata support.
You can change the serial port name in the text field.  Clicking the little reload arrow in the bottom right of the pane will re-initialize Firmata.  'start' is the first pin # (starting at 0) that is being used for the array of sensors.  Currently stores the values into a compound parameter ranging from 0 to 1.  For real analog values, the result should be divided  by 1023 (per arduino analog pin specs).