DanceTileEfx.
This is like a mask effect that masks out everything except activated tiles.  This is a first hack.  It is not ideal because we are always rendering the channel.  We could use our loop in UIModeSelector to enable/disable this channel depending on whether any tile is active.  Also patterns should render with a background alpha of 0.   The channel blend mode is Normal so this pattern will decal over anything else running.